+++ docs-util/oas-output/operations/admin/post_admin_inventory-items.ts
@@ -23,8 +23,6 @@
  * tags:
  *   - Inventory Items
  * responses:
- *   "200":
- *     description: OK
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "401":