+++ docs-util/oas-output/operations/admin/post_admin_products_[id]_options_[option_id].ts
@@ -30,8 +30,6 @@
  * tags:
  *   - Products
  * responses:
- *   "200":
- *     description: OK
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "401":
@@ -44,6 +42,10 @@
  *     $ref: "#/components/responses/invalid_request_error"
  *   "500":
  *     $ref: "#/components/responses/500_error"
+ * requestBody:
+ *   content:
+ *     application/json:
+ *       schema: {}
  * 
 */