+++ packages/medusa/src/api/routes/admin/price-lists/create-price-list.ts
@@ -11,6 +11,7 @@ import {
 import { Request } from "express"
 import { EntityManager } from "typeorm"
 import { defaultAdminPriceListFields, defaultAdminPriceListRelations } from "."
+import { featureFlagRouter } from "../../../../loaders/feature-flags"
 import TaxInclusivePricingFeatureFlag from "../../../../loaders/feature-flags/tax-inclusive-pricing"
 import { PriceList } from "../../../../models"
 import PriceListService from "../../../../services/price-list"