+++ packages/types/src/pricing/service.ts
@@ -10,15 +10,15 @@ import {
   CreateRuleTypeDTO,
   FilterablePriceListProps,
   FilterablePriceListRuleProps,
+  FilterablePriceProps,
   FilterablePriceRuleProps,
-  FilterablePriceSetMoneyAmountProps,
   FilterablePriceSetProps,
   FilterableRuleTypeProps,
+  PriceDTO,
   PriceListDTO,
   PriceListRuleDTO,
   PriceRuleDTO,
   PriceSetDTO,
-  PriceSetMoneyAmountDTO,
   PricingContext,
   PricingFilters,
   RemovePriceListRulesDTO,
@@ -169,7 +169,7 @@ export interface IPricingModuleService extends IModuleService {
    *   const priceSet = await pricingService.retrieve(
    *     priceSetId,
    *     {
-   *       relations: ["money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -230,7 +230,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceSetIds
    *     },
    *     {
-   *       relations: ["money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -253,7 +253,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceSetIds
    *     },
    *     {
-   *       relations: ["money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -270,7 +270,7 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSets (priceSetIds: string[], moneyAmountIds: string[], skip: number, take: number) {
+   * async function retrievePriceSets (priceSetIds: string[], priceIds: string[], skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
    *   const priceSets = await pricingService.list(
@@ -280,14 +280,14 @@ export interface IPricingModuleService extends IModuleService {
    *           id: priceSetIds
    *         },
    *         {
-   *           money_amounts: {
-   *             id: moneyAmountIds
+   *           prices: {
+   *             id: priceIds
    *           }
    *         }
    *       ]
    *     },
    *     {
-   *       relations: ["money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -350,7 +350,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceSetIds
    *     },
    *     {
-   *       relations: ["money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -373,7 +373,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceSetIds
    *     },
    *     {
-   *       relations: ["money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -390,7 +390,7 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSets (priceSetIds: string[], moneyAmountIds: string[], skip: number, take: number) {
+   * async function retrievePriceSets (priceSetIds: string[], priceIds: string[], skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
    *   const [priceSets, count] = await pricingService.listAndCount(
@@ -400,14 +400,14 @@ export interface IPricingModuleService extends IModuleService {
    *           id: priceSetIds
    *         },
    *         {
-   *           money_amounts: {
-   *             id: moneyAmountIds
+   *           prices: {
+   *             id: priceIds
    *           }
    *         }
    *       ]
    *     },
    *     {
-   *       relations: ["money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -1240,52 +1240,52 @@ export interface IPricingModuleService extends IModuleService {
   deleteRuleTypes(ruleTypeIds: string[], sharedContext?: Context): Promise<void>
 
   /**
-   * This method is used to retrieve a paginated list of price set money amounts based on optional filters and configuration.
+   * This method is used to retrieve a paginated list of prices based on optional filters and configuration.
    *
-   * @param {FilterablePriceSetMoneyAmountProps} filters - The filters to apply on the retrieved price set money amounts.
-   * @param {FindConfig<PriceSetMoneyAmountDTO>} config -
-   * The configurations determining how the price set money amounts are retrieved. Its properties, such as `select` or `relations`, accept the
-   * attributes or relations associated with a price set money amount.
+   * @param {FilterablePriceProps} filters - The filters to apply on the retrieved prices.
+   * @param {FindConfig<PriceDTO>} config -
+   * The configurations determining how the prices are retrieved. Its properties, such as `select` or `relations`, accept the
+   * attributes or relations associated with a price.
    * @param {Context} sharedContext - A context used to share resources, such as transaction manager, between the application and the module.
-   * @returns {Promise<PriceSetMoneyAmountDTO[]>} The list of price set money amounts.
+   * @returns {Promise<PriceDTO[]>} The list of prices.
    *
    * @example
    *
-   * To retrieve a list of price set money amounts using their IDs:
+   * To retrieve a list of prices using their IDs:
    *
    * ```ts
    * import {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string) {
+   * async function retrievePrices (id: string) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const priceSetMoneyAmounts = await pricingService.listPriceSetMoneyAmounts({
+   *   const prices = await pricingService.listPrices({
    *     id: [id]
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
-   * To specify relations that should be retrieved within the price set money amounts:
+   * To specify relations that should be retrieved within the prices:
    *
    * ```ts
    * import {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string) {
+   * async function retrievePrices (id: string) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const priceSetMoneyAmounts = await pricingService.listPriceSetMoneyAmounts({
+   *   const prices = await pricingService.listPrices({
    *     id: [id]
    *   }, {
    *     relations: ["price_rules"]
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
@@ -1296,10 +1296,10 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string, skip: number, take: number) {
+   * async function retrievePrices (id: string, skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const priceSetMoneyAmounts = await pricingService.listPriceSetMoneyAmounts({
+   *   const prices = await pricingService.listPrices({
    *     id: [id]
    *   }, {
    *     relations: ["price_rules"],
@@ -1307,7 +1307,7 @@ export interface IPricingModuleService extends IModuleService {
    *     take
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
@@ -1318,10 +1318,10 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (ids: string[], titles: string[], skip: number, take: number) {
+   * async function retrievePrices (ids: string[], titles: string[], skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const priceSetMoneyAmounts = await pricingService.listPriceSetMoneyAmounts({
+   *   const prices = await pricingService.listPrices({
    *     $and: [
    *       {
    *         id: ids
@@ -1336,78 +1336,76 @@ export interface IPricingModuleService extends IModuleService {
    *     take
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    */
-  listPriceSetMoneyAmounts(
-    filters?: FilterablePriceSetMoneyAmountProps,
-    config?: FindConfig<PriceSetMoneyAmountDTO>,
+  listPrices(
+    filters?: FilterablePriceProps,
+    config?: FindConfig<PriceDTO>,
     sharedContext?: Context
-  ): Promise<PriceSetMoneyAmountDTO[]>
+  ): Promise<PriceDTO[]>
 
-  softDeletePriceSetMoneyAmounts<
-    TReturnableLinkableKeys extends string = string
-  >(
-    psmaIds: string[],
+  softDeletePrices<TReturnableLinkableKeys extends string = string>(
+    priceIds: string[],
     config?: SoftDeleteReturn<TReturnableLinkableKeys>,
     sharedContext?: Context
   ): Promise<Record<string, string[]> | void>
 
-  restorePriceSetMoneyAmounts<TReturnableLinkableKeys extends string = string>(
-    psmaIds: string[],
+  restorePrices<TReturnableLinkableKeys extends string = string>(
+    priceIds: string[],
     config?: RestoreReturn<TReturnableLinkableKeys>,
     sharedContext?: Context
   ): Promise<Record<string, string[]> | void>
 
   /**
-   * This method is used to retrieve a paginated list of price set money amounts along with the total count of
-   * available price set money amounts satisfying the provided filters.
+   * This method is used to retrieve a paginated list of prices along with the total count of
+   * available prices satisfying the provided filters.
    *
-   * @param {FilterablePriceSetMoneyAmountProps} filters - The filters to apply on the retrieved price set money amounts.
-   * @param {FindConfig<PriceSetMoneyAmountDTO>} config -
-   * The configurations determining how the price set money amounts are retrieved. Its properties, such as `select` or `relations`, accept the
-   * attributes or relations associated with a price set money amount.
+   * @param {FilterablePriceProps} filters - The filters to apply on the retrieved prices.
+   * @param {FindConfig<PriceDTO>} config -
+   * The configurations determining how the prices are retrieved. Its properties, such as `select` or `relations`, accept the
+   * attributes or relations associated with a price.
    * @param {Context} sharedContext - A context used to share resources, such as transaction manager, between the application and the module.
-   * @returns {Promise<[PriceSetMoneyAmountDTO[], number]>} The list of price set money amounts and their total count.
+   * @returns {Promise<[PriceDTO[], number]>} The list of prices and their total count.
    *
    * @example
    *
-   * To retrieve a list of price set money amounts using their IDs:
+   * To retrieve a list of prices using their IDs:
    *
    * ```ts
    * import {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string) {
+   * async function retrievePrices (id: string) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const [priceSetMoneyAmounts, count] = await pricingService.listAndCountPriceSetMoneyAmounts({
+   *   const [prices, count] = await pricingService.listAndCountPrices({
    *     id: [id]
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
-   * To specify relations that should be retrieved within the price set money amounts:
+   * To specify relations that should be retrieved within the prices:
    *
    * ```ts
    * import {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string) {
+   * async function retrievePrices (id: string) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const [priceSetMoneyAmounts, count] = await pricingService.listAndCountPriceSetMoneyAmounts({
+   *   const [prices, count] = await pricingService.listAndCountPrices({
    *     id: [id]
    *   }, {
    *     relations: ["price_rules"],
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
@@ -1418,10 +1416,10 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (id: string, skip: number, take: number) {
+   * async function retrievePrices (id: string, skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const [priceSetMoneyAmounts, count] = await pricingService.listAndCountPriceSetMoneyAmounts({
+   *   const [prices, count] = await pricingService.listAndCountPrices({
    *     id: [id]
    *   }, {
    *     relations: ["price_rules"],
@@ -1429,7 +1427,7 @@ export interface IPricingModuleService extends IModuleService {
    *     take
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    *
@@ -1440,10 +1438,10 @@ export interface IPricingModuleService extends IModuleService {
    *   initialize as initializePricingModule,
    * } from "@medusajs/pricing"
    *
-   * async function retrievePriceSetMoneyAmounts (ids: string[], titles: string[], skip: number, take: number) {
+   * async function retrievePrices (ids: string[], titles: string[], skip: number, take: number) {
    *   const pricingService = await initializePricingModule()
    *
-   *   const [priceSetMoneyAmounts, count] = await pricingService.listAndCountPriceSetMoneyAmounts({
+   *   const [prices, count] = await pricingService.listAndCountPrices({
    *     $and: [
    *       {
    *         id: ids
@@ -1458,15 +1456,15 @@ export interface IPricingModuleService extends IModuleService {
    *     take
    *   })
    *
-   *   // do something with the price set money amounts or return them
+   *   // do something with the prices or return them
    * }
    * ```
    */
-  listAndCountPriceSetMoneyAmounts(
-    filters?: FilterablePriceSetMoneyAmountProps,
-    config?: FindConfig<PriceSetMoneyAmountDTO>,
+  listAndCountPrices(
+    filters?: FilterablePriceProps,
+    config?: FindConfig<PriceDTO>,
     sharedContext?: Context
-  ): Promise<[PriceSetMoneyAmountDTO[], number]>
+  ): Promise<[PriceDTO[], number]>
 
   /**
    * This method is used to retrieve a price rule by its ID.
@@ -1750,7 +1748,7 @@ export interface IPricingModuleService extends IModuleService {
    *   priceSetId: string,
    *   ruleTypeId: string,
    *   value: string,
-   *   priceSetMoneyAmountId: string,
+   *   priceId: string,
    *   priceListId: string
    * ) {
    *   const pricingService = await initializePricingModule()
@@ -1761,7 +1759,7 @@ export interface IPricingModuleService extends IModuleService {
    *       price_set_id: priceSetId,
    *       rule_type_id: ruleTypeId,
    *       value,
-   *       price_set_money_amount_id: priceSetMoneyAmountId,
+   *       price_id: priceId,
    *       price_list_id: priceListId
    *     }
    *   ])
@@ -1874,7 +1872,7 @@ export interface IPricingModuleService extends IModuleService {
    *   const priceList = await pricingService.retrievePriceList(
    *     priceListId,
    *     {
-   *       relations: ["price_set_money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -1935,7 +1933,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceListIds
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -1958,7 +1956,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceListIds
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -1990,7 +1988,7 @@ export interface IPricingModuleService extends IModuleService {
    *       ]
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -2053,7 +2051,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceListIds
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"]
+   *       relations: ["prices"]
    *     }
    *   )
    *
@@ -2076,7 +2074,7 @@ export interface IPricingModuleService extends IModuleService {
    *       id: priceListIds
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }
@@ -2108,7 +2106,7 @@ export interface IPricingModuleService extends IModuleService {
    *       ]
    *     },
    *     {
-   *       relations: ["price_set_money_amounts"],
+   *       relations: ["prices"],
    *       skip,
    *       take
    *     }