+++ www/apps/api-reference/app/api/download/[area]/route.ts
@@ -1,6 +1,7 @@
 import { existsSync, readFileSync } from "fs"
 import { NextResponse } from "next/server"
 import path from "path"
+import { Version } from "../../../../types/openapi"
 
 type DownloadParams = {
   params: {
@@ -9,8 +10,16 @@ type DownloadParams = {
 }
 
 export function GET(request: Request, { params }: DownloadParams) {
+  const { searchParams } = new URL(request.url)
   const { area } = params
-  const filePath = path.join(process.cwd(), `specs/${area}/openapi.full.yaml`)
+  const version =
+    process.env.NEXT_PUBLIC_VERSIONING === "true"
+      ? (searchParams.get("version") as Version) || "1"
+      : "1"
+  const filePath = path.join(
+    process.cwd(),
+    `${version === "1" ? "specs" : "specs-v2"}/${area}/openapi.full.yaml`
+  )
 
   if (!existsSync(filePath)) {
     return new NextResponse(null, {